+++ Tracker.Core/Models/Fees/HybridSubcharge.cs
@@ -34,7 +34,7 @@ namespace Tracker.Core.Models.Fees
             _flatCost = flatCost;
         }
 
-        public bool CheckValue(int value)
+        public bool CheckValue(decimal value)
         {
             if (value >= Minimum && value <= Maximum)
             {
@@ -43,7 +43,7 @@ namespace Tracker.Core.Models.Fees
             return false;
         }
 
-        public decimal GetCost(int count)
+        public decimal GetCost(decimal count)
         {
             return (Cost * ((count - Minimum)/_groupSize)) + _flatCost;
         }
