+++ Tracker.Core/Models/Fees/ISubcharge.cs
@@ -5,7 +5,7 @@
         int Minimum { get; }
         int Maximum { get; }
         decimal Cost { get; }
-        bool CheckValue(int value);
-        decimal GetCost(int count);
+        bool CheckValue(decimal value);
+        decimal GetCost(decimal count);
     }
 }
