+++ Tracker.Core/Models/Fees/VariableCategorySubcharge.cs
@@ -32,7 +32,7 @@ namespace Tracker.Core.Models.Fees
             _flatCost = flatCost;
         }
 
-        public bool CheckValue(int value)
+        public bool CheckValue(decimal value)
         {
             if (value >= Minimum && value <= Maximum)
             {
@@ -41,7 +41,7 @@ namespace Tracker.Core.Models.Fees
             return false;
         }
 
-        public decimal GetCost(int count)
+        public decimal GetCost(decimal count)
         {
             return (Cost * (count - Minimum + 1)) + _flatCost;
         }
